Calculate price - refactor: * removed not needed classes * moved classes to new packages * simplified Retrieval classes * added more tests

+++ src/main/kotlin/com/zen/recruitment/discount/domain/PercentageDiscount.kt
@@ -6,5 +6,5 @@ import jakarta.persistence.Entity
 @Entity
 @DiscriminatorValue("PERCENTAGE")
 class PercentageDiscount(
-    val percentage: Double
+    val percentage: Int
 ) : Discount()
